move main launchpad logic into 'core' pkg
