+++ packages/core/test/config.js
