+++ packages/core/test/test.bat
